Import GRUB code to embed blocklist into core.img

+++ src/endless/grub/grub.h
@@ -1,11 +1,12 @@
 #pragma once
 
 #include <Windows.h>
+#include <inttypes.h>
 
 
-typedef unsigned short grub_uint16_t;
-typedef unsigned int grub_uint32_t;
-typedef unsigned long long grub_uint64_t;
+typedef uint16_t grub_uint16_t;
+typedef uint32_t grub_uint32_t;
+typedef uint64_t grub_uint64_t;
 
 typedef grub_uint64_t grub_size_t;
 typedef int grub_err_t;
